completely refactor the seeding and make it way more maintanable
