+++ db/seeder/config/seed_config.rb
@@ -1,0 +1,51 @@
+module SeedConfig
+  RANGE_OF_STUDIES = 3
+  RANGE_OF_EXPERIENCES = 3
+  RANGE_OF_APPLICATIONS = 5
+
+  SALARY_RANGE = 30000..60000
+
+  THREADS_TO_USE = 4
+  IMAGE_BATCH_SIZE = 20
+
+  TEMP_IMG_PATH = 'app/assets/images/tmp_images/'
+
+  REAL_CITIES = [
+    "Paris, France",
+    "New York, NY, United States",
+    "Brussels, Belgium",
+    "London, UK",
+    "Rome, Italy"
+  ]
+
+  REAL_LANGUAGES = [
+    "English",
+    "French",
+    "Spanish",
+    "German",
+    "Italian",
+    "Chinese",
+    "Japanese",
+    "Russian",
+    "Arabic",
+    "Portuguese",
+    "Dutch",
+    "Korean",
+    "Turkish",
+    "Polish",
+    "Swedish",
+    "Danish",
+    "Norwegian",
+    "Finnish",
+    "Greek",
+    "Czech",
+    "Hungarian",
+    "Romanian",
+    "Bulgarian",
+    "Croatian",
+    "Slovak",
+    "Slovenian",
+    "Lithuanian",
+    "Latvian",
+    "Estonian"]
+end
